properly modularising rocon_gateway and rocon_hub_client
