+++ rocon_gateway/src/rocon_gateway/__init__.py
@@ -22,7 +22,4 @@ from .gateway_node import GatewayNode
 from .graph import Graph
 import gateway
 from .gateway import Gateway
-import hub_api
-import hub_discovery
-import hub_manager
 import samples
