+++ rocon_gateway/src/rocon_gateway/watcher_thread.py
@@ -11,8 +11,6 @@ import rospy
 import threading
 import httplib
 
-from .exceptions import HubConnectionLostError
-
 ##############################################################################
 # Watcher
 ##############################################################################
