+++ rocon_hub_client/src/rocon_hub_client/__init__.py
@@ -1,1 +1,13 @@
+#!/usr/bin/env python
+#
+# License: BSD
+#   https://raw.github.com/robotics-in-concert/rocon_multimaster/hydro-devel/rocon_hub_client/LICENSE
+#
 
+import hub_api
+from .hub_client import Hub
+from .hub_discovery import HubDiscovery
+from .exceptions import HubError, \
+                        HubNotFoundError, HubNameNotFoundError, \
+                        HubConnectionBlacklistedError, HubConnectionNotWhitelistedError, \
+                        HubConnectionAlreadyExistsError, HubConnectionLostError
