+++ rocon_hub_client/src/rocon_hub_client/hub_client.py
@@ -3,75 +3,60 @@
 # License: BSD
 #   https://raw.github.com/robotics-in-concert/rocon_multimaster/master/rocon_hub_client/LICENSE
 #
-##############################################################################
+###############################################################################
 # Imports
+###############################################################################
+
+import redis
+import rospy
+
+# local imports
+import hub_api
+from .exceptions import HubNameNotFoundError, HubNotFoundError, \
+                        HubConnectionBlacklistedError, HubConnectionNotWhitelistedError
+
+##############################################################################
+# Hub
 ##############################################################################
 
-import rospy
-import redis
-from .hub_connector import HubConnector
 
+class Hub(object):
 
-class HubClient(HubConnector):
+    def __init__(self, ip, port, whitelist=[], blacklist=[]):
+        '''
+          @param remote_gateway_request_callbacks : to handle redis responses
+          @type list of function pointers (back to GatewaySync class
 
-    # HubConnector provide
-    #    connect(hub_uri)
+          @param ip : redis server ip
+          @param port : redis server port
 
-    ConnectionError = redis.exceptions.ConnectionError
-
-    pool = None
-    server = None
-    pubsub = None
-    callback = None
-    namespace = ''
-    param = {}
-
-    def __init__(self, whitelist, blacklist, is_zeroconf, namespace, name, callbacks):
-        super(HubClient, self).__init__(whitelist, blacklist, is_zeroconf, self._connectToHub)
-
-        self._namespace = namespace
-        self._name = name
-        self._unique_name = ''
-        self._callbacks = callbacks
-
+          @raise HubNameNotFoundError, HubNotFoundError
+        '''
+        # variables
+        self.uri = str(ip) + ":" + str(port)
         self._redis_keys = {}
-#        self._redis_keys['index'] = createHubKey('index')
         self._redis_channels = {}
-        self._redis_pubsub_server = None
 
-    def _connectToHub(self, ip, portarg):
+        # redis server connection
         try:
-            self.pool = redis.ConnectionPool(host=ip, port=portarg, db=0)
-            self.server = redis.Redis(connection_pool=self.pool)
-            rospy.logdebug("HubClient : connected to the hub's redist server.")
-            self._redis_pubsub_server = self.server.pubsub()
-        except redis.exceptions.ConnectionError as unused_e:
-            rospy.logerror("HubClient : failed rule to the hub's redis server.")
-            return False
-            raise
-        return True
-
-    def registerKey(self, key, element):
-        key = self._namespace + ":" + key
-        element = self._namespace + ":" + element
-
-        if self.server.sadd(key, element):
-            return True
-        else:
-            raise Exception("Same concertmaster name exist")
-            return False
-
-    def unregisterKey(self, key, element):
-        try:
-            key = self._namespace + ":" + key
-            element = self._namespace + ":" + element
-            self.server.srem(key, element)
+            self.pool = redis.ConnectionPool(host=ip, port=port, db=0)
+            self._redis_server = redis.Redis(connection_pool=self.pool)
+            self._redis_pubsub_server = self._redis_server.pubsub()
+            hub_key_name = self._redis_server.get("rocon:hub:name")
+            # Be careful, hub_name is None, it means the redis server is
+            # found but hub_name not yet set or not set at all.
+            if not hub_key_name:
+                self._redis_server = None
+                raise HubNameNotFoundError("couldn't resolve hub name on the redis server [%s:%s]" % (ip, port))
+            else:
+                self.name = hub_api.key_base_name(hub_key_name)  # perhaps should store all key names somewhere central
+                rospy.logdebug("Gateway : resolved hub name [%s].", self.name)
         except redis.exceptions.ConnectionError:
-            raise  # ~HubClient.ConnectionError
+            self._redis_server = None
+            raise HubNotFoundError("couldn't connect to the redis server")
 
-    def getValues(self, key):
-        key = self._namespace + ":" + key
-        values = self.server.smembers(key)
-
-        values = [v[len(self._namespace) + 1:] for v in values]
-        return values
+        # whitelists, blacklists
+        if self.uri in blacklist or self.name in blacklist:
+            raise HubConnectionBlacklistedError("ignoring blacklisted hub [%s]" % self.uri)
+        if not ((len(whitelist) == 0) or (self.uri in whitelist) or (self.name in whitelist)):
+            raise HubConnectionNotWhitelistedError("hub/ip not in non-empty whitelist [%s]%s" % (self.name, whitelist))
