+++ rocon_hub_client/src/rocon_hub_client/hub_discovery.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 #
 # License: BSD
-#   https://raw.github.com/robotics-in-concert/rocon_multimaster/hydro-devel/rocon_gateway/LICENSE
+#   https://raw.github.com/robotics-in-concert/rocon_multimaster/hydro-devel/rocon_hub_client/LICENSE
 #
 
 ###############################################################################
